+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Updated image cdns.
 - Added Eidolon#picture, CombatType#icon, and Path#icon.
 - Added "PS" to type of StarRailUser#platform.
+- Added StarRailUser#playStationAccount.
 # 1.0.1
 - Added StarRailCharacterBuild#imageUrl.
 - Removed SkillTreeNode#previousNodeIds, please use SkillTreeNode#previousNodeId instead.